requirements.txt
Flask==2.3.2
flask-cors==4.0.0
pymongo==4.5.0
gunicorn==21.2.0
requests==2.28.1
mysql-connector-python==8.0.32

